cloud/functions/brain_loop/requirements.txt
functions-framework==3.*
yfinance>=0.2.31
google-cloud-storage>=2.14.0
feedparser>=6.0.0
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.10.0
